+++ aula dia 01-08-2023/pyside exer 6.py
@@ -22,7 +22,7 @@ class MainWindow(QMainWindow):
         self.button.clicked.connect(self.calcular_metros)
 
     def calcular_metros(self):
-        num1 = int(self.input1.text())
+        num1 = float(self.input1.text())
         soma = num1*100
         self.result_label.setText(f"converter metros para centímetros: {soma}")
